.: FEATURE: Added support for 64 bit architecture

+++ NSStringAdditions.h
@@ -26,7 +26,7 @@
 
 + (NSString *)stringByGeneratingUUID;
 
-+ (NSString *)base64StringFromData: (NSData *)data length: (int)length;
++ (NSString *)base64StringFromData: (NSData *)data length: (NSUInteger)length;
 
 #pragma mark -
