+++ XMLRPCConnectionManager.m
@@ -81,7 +81,7 @@ static XMLRPCConnectionManager *sharedInstance = nil;
     return [myConnections allKeys];
 }
 
-- (int)numberOfActiveConnections {
+- (NSUInteger)numberOfActiveConnections {
     return [myConnections count];
 }
